requirements.txt
requests
beautifulsoup4
LXML
PyMuPDF
streamlit
pytesseract
Pillow
PyMuPDF
google-generativeai
gtts
sentence-transformers
numpy
tqdm
